requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Kafka
kafka-python==2.0.2

# HTTP Client
httpx==0.26.0

# File handling
python-magic==0.4.27

# Monitoring
prometheus-client==0.19.0

# OpenAI-compatible client (for Ollama)
openai==1.10.0
